Create Lip Gather-APP Project
We developed a mobile app specifically for collecting datasets. It includes comprehensive automated guidance to make data collection more convenient for participants. Before starting the recording, we explained the complete data collection precautions to the participants, such as ensuring recordings are made in a quiet environment.

Additionally, we instructed participants to ensure their face is within the video frame, directly facing the camera, and to avoid having other people's faces within the recording area. Participants are required to hold the phone with one hand during recording and maintain an appropriate distance from the front camera to obtain clear and properly framed video images. To prevent mobile notification pop-ups from covering the recording interface's prompt words, it is recommended to turn off message notification functions for apps like WeChat.

We introduced the specific recording pipeline in advance, such as each set containing 30 sentences, to encourage regular breaks during the task. We suggested participants take appropriate rest after completing a certain number of recordings to maintain a good experience and performance. The recording duration for each sentence is 3 seconds, after which the system automatically navigates to the recording interface for the next sentence. If errors are discovered or recording is interrupted, participants are guided to return to the main menu to restart recording.

Video recording is done using the phone's front camera, with the video resolution set to 1920x1080 and the standard frame rate set at 30 frames per second. To enhance the stability of the data collection app and reduce the data upload burden, we set up an encrypted cloud server for data transmission and ultimately download the data to a local server for storage and processing.